✨ RedirectAuth 페이지에서 요청한 API의 에러코드에 따른 에러 컴포넌트를 렌더링한다. - status가 400이면 중복되는 이메일을 요청한 것이므로 DuplicateEmail 컴포넌트를 렌더링한다. - 이외의 status는 에러 메시지를 화면에 렌더링 한다.
