+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import Home from '@/pages/Home';
 import NotFound from '@/pages/NotFound';
 
-import RedirectAuth from '@/pages/Public/RedirectAuth';
+import RedirectAuth from '@/pages/Public/RedirectAuth/RedirectAuth';
 import Login from '@/pages/Public/Login';
 import OAuthSignUp from '@/pages/Public/SignUp-OAuth';
 import CommonSignUp from '@/pages/Public/SignUp-Common';
